fixed bugs in conversion specifiers

+++ write_hex_bigLetter.c
@@ -14,7 +14,7 @@ size_t write_hex_bigLetter(va_list args, flag_t *flag,
 	size_t nwrite = 0;
 	uintmax_t num = (length_modifier == SHORT)
 		? va_arg(args, unsigned int)
-		: va_arg(args, unsigned long int);
+		: va_arg(args, unsigned int);
 
 	if (flag->hash && num > 0)
 		nwrite += _puts("0X");
